requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
pandas==2.1.3
python-dotenv==1.0.0
openai==1.3.7
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
beautifulsoup4==4.12.2
lxml==4.9.3
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

